requirements.txt
beautifulsoup4==4.5.0
dj-database-url==0.4.1
Django==1.9.8
django-cors-headers==1.1.0
django-filter==0.13.0
djangorestframework==3.4.0
djangorestframework-jsonapi==2.0.1
djangorestframework-jwt==1.8.0
gunicorn==19.6.0
httplib2==0.9.2
inflection==0.3.1
nflgame==1.2.19
psycopg2==2.6.2
PyJWT==1.4.1
pytz==2016.6.1
whitenoise==3.2
